Move common capabilities to OrdinaryUser
