+++ src/main/kotlin/com/terraformation/backend/customer/model/OrdinaryUser.kt
@@ -1,6 +1,7 @@
 package com.terraformation.backend.customer.model
 
 import com.terraformation.backend.auth.CurrentUserHolder
+import com.terraformation.backend.auth.SuperAdminAuthority
 import com.terraformation.backend.auth.currentUser
 import com.terraformation.backend.customer.db.ParentStore
 import com.terraformation.backend.customer.db.PermissionStore
@@ -116,7 +117,11 @@ abstract class OrdinaryUser(
     get() = makeFullName(firstName, lastName)
 
   override fun getAuthorities(): MutableCollection<out GrantedAuthority> {
-    return mutableSetOf()
+    return if (isSuperAdmin()) {
+      mutableSetOf(SuperAdminAuthority)
+    } else {
+      mutableSetOf()
+    }
   }
 
   override fun getPassword(): String {
@@ -689,31 +694,123 @@ abstract class OrdinaryUser(
 
   // Capabilities
 
-  protected open fun isAcceleratorAdmin() = false
+  /** Returns true if the user is an admin, owner or Terraformation Contact of any organizations. */
+  override fun hasAnyAdminRole() =
+      organizationRoles.values.any {
+        it == Role.Owner || it == Role.Admin || it == Role.TerraformationContact
+      }
 
-  protected open fun isTFExpertOrHigher() = false
+  override fun isAdminOrHigher(organizationId: OrganizationId?): Boolean {
+    return organizationId?.let {
+      recordPermissionCheck(RolePermissionCheck(Role.Admin, organizationId))
+      when (organizationRoles[organizationId]) {
+        Role.Admin,
+        Role.Owner,
+        Role.TerraformationContact -> true
+        else -> false
+      }
+    } ?: false
+  }
 
-  protected open fun isReadOnlyOrHigher() = false
+  private fun isAcceleratorAdmin(): Boolean {
+    recordPermissionCheck(GlobalRolePermissionCheck(GlobalRole.AcceleratorAdmin))
+    return setOf(GlobalRole.AcceleratorAdmin, GlobalRole.SuperAdmin).any { it in globalRoles }
+  }
 
-  protected open fun isSuperAdmin() = false
+  private fun isTFExpertOrHigher(): Boolean {
+    recordPermissionCheck(GlobalRolePermissionCheck(GlobalRole.TFExpert))
+    return setOf(GlobalRole.TFExpert, GlobalRole.AcceleratorAdmin, GlobalRole.SuperAdmin).any {
+      it in globalRoles
+    }
+  }
 
-  protected open fun isOwner(organizationId: OrganizationId?) = false
+  private fun isReadOnlyOrHigher(): Boolean {
+    recordPermissionCheck(GlobalRolePermissionCheck(GlobalRole.ReadOnly))
+    return setOf(
+            GlobalRole.ReadOnly,
+            GlobalRole.TFExpert,
+            GlobalRole.AcceleratorAdmin,
+            GlobalRole.SuperAdmin)
+        .any { it in globalRoles }
+  }
 
-  protected open fun isAdminOrHigher(facilityId: FacilityId?) = false
+  private fun isSuperAdmin(): Boolean {
+    recordPermissionCheck(GlobalRolePermissionCheck(GlobalRole.SuperAdmin))
+    return GlobalRole.SuperAdmin in globalRoles
+  }
 
-  protected open fun isManagerOrHigher(organizationId: OrganizationId?) = false
+  private fun isOwner(organizationId: OrganizationId?) =
+      organizationId?.let {
+        recordPermissionCheck(RolePermissionCheck(Role.Owner, organizationId))
+        organizationRoles[organizationId] == Role.Owner
+      } ?: false
 
-  protected open fun isManagerOrHigher(facilityId: FacilityId?) = false
+  private fun isAdminOrHigher(facilityId: FacilityId?) =
+      facilityId?.let {
+        recordPermissionCheck(RolePermissionCheck(Role.Admin, facilityId))
+        when (facilityRoles[facilityId]) {
+          Role.Admin,
+          Role.Owner,
+          Role.TerraformationContact -> true
+          else -> false
+        }
+      } ?: false
 
-  protected open fun isManagerOrHigher(plantingSiteId: PlantingSiteId?) = false
+  private fun isManagerOrHigher(organizationId: OrganizationId?) =
+      organizationId?.let {
+        recordPermissionCheck(RolePermissionCheck(Role.Manager, organizationId))
+        when (organizationRoles[organizationId]) {
+          Role.Admin,
+          Role.Manager,
+          Role.Owner,
+          Role.TerraformationContact -> true
+          else -> false
+        }
+      } ?: false
 
-  protected open fun isMember(facilityId: FacilityId?) = false
+  private fun isManagerOrHigher(facilityId: FacilityId?) =
+      facilityId?.let {
+        recordPermissionCheck(RolePermissionCheck(Role.Manager, facilityId))
+        when (facilityRoles[facilityId]) {
+          Role.Admin,
+          Role.Manager,
+          Role.Owner,
+          Role.TerraformationContact -> true
+          else -> false
+        }
+      } ?: false
 
-  protected open fun isMember(organizationId: OrganizationId?) = false
+  private fun isManagerOrHigher(plantingSiteId: PlantingSiteId?) =
+      plantingSiteId?.let {
+        recordPermissionCheck(RolePermissionCheck(Role.Manager, plantingSiteId))
+        isManagerOrHigher(parentStore.getOrganizationId(plantingSiteId))
+      } ?: false
 
-  protected open fun isGlobalReader(organizationId: OrganizationId) = false
+  private fun isMember(facilityId: FacilityId?) =
+      facilityId?.let {
+        recordPermissionCheck(RolePermissionCheck(Role.Contributor, facilityId))
+        facilityId in facilityRoles
+      } ?: false
 
-  protected open fun isGlobalWriter(organizationId: OrganizationId) = false
+  private fun isMember(organizationId: OrganizationId?) =
+      organizationId?.let {
+        recordPermissionCheck(RolePermissionCheck(Role.Contributor, organizationId))
+        organizationId in organizationRoles
+      } ?: false
+
+  /** Returns true if one of the user's global roles allows them to read an organization. */
+  private fun isGlobalReader(organizationId: OrganizationId) =
+      GlobalRole.SuperAdmin in globalRoles ||
+          (isReadOnlyOrHigher() &&
+              (parentStore.hasInternalTag(organizationId, InternalTagIds.Accelerator) ||
+                  parentStore.hasApplications(organizationId)))
+
+  /** Returns true if one of the user's global roles allows them to write to an organization. */
+  private fun isGlobalWriter(organizationId: OrganizationId) =
+      GlobalRole.SuperAdmin in globalRoles ||
+          (isTFExpertOrHigher() &&
+              (parentStore.hasInternalTag(organizationId, InternalTagIds.Accelerator) ||
+                  parentStore.hasApplications(organizationId)))
 
   private var isRecordingChecks: Boolean = false
